+++ src/com/intellij/idea/plugin/hybris/common/utils/HybrisIcons.kt
@@ -90,7 +90,7 @@ object HybrisIcons {
     val EXTENSION_BACKOFFICE = AllIcons.Nodes.Folder
     val EXTENSION_HMC = AllIcons.Nodes.Folder
     val EXTENSION_HAC = AllIcons.Nodes.Folder
-    val EXTENSION_KOTLIN_NATURE = KotlinIcons.SMALL_LOGO
+    val EXTENSION_KOTLIN_NATURE = if (PluginCommon.isPluginActive(PluginCommon.PLUGIN_KOTLIN)) KotlinIcons.SMALL_LOGO else AllIcons.Nodes.Module
 
     val TYPE_SYSTEM = getIcon("/icons/typeSystem.svg")
     val SPRING_BEAN = getIcon("icons/springBean.svg")