+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 
 ### Fixes
 - Configure SAP CX JavaDoc for OOTB Modules [#1027](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1027) for Issue [1026](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1026)
+- Removed dependency on Kotlin PSI utils methods [#1053](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1053) for Issue [1052](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1052)
 
 ### Other
 - Added JetBrains Marketplace error reporting [#1039](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1027) for Issue [1026](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1039)